refactor: homework 서비스 인터페이스 제거 - 구현체가 하나뿐이라 추상화 이득 없음
AssignmentService / SubmissionService 인터페이스를 삭제하고
XxxServiceImpl 을 XxxService 로 정리했다.

인터페이스가 계약이 아니라 구현을 담고 있었다.
AssignmentService 145줄 중 대부분이 DTO 매핑 default 메서드였고,
실제 계약은 메서드 시그니처 10개뿐이었다.

  default StudentSubmissionStatusDTO statusToDto(...)   { ... }
  default AssignmentDTO domainToDto(...)                { ... }
  default Assignment createDtoToDomain(...)             { ... }

매핑 메서드는 서비스 클래스로 옮기고 가시성(public)을 그대로 유지했다.
인터페이스의 default 메서드는 public 이므로 동작이 바뀌지 않는다.
매핑 로직을 서비스가 들고 있는 것은 별개 문제이며 후속으로 다룬다.

측정
  homework  28파일 2,634줄 -> 26파일 2,459줄
  변경      12파일 +130 -309 (순감소 179줄)
  테스트    171건 전부 통과 (동작 불변)

See also: #3
